Move RegistrationValidationView into user_authn.
